(tox) Fixed template name in failing test
The current version of dbtemplates states that the Template model name field must be unique (this constraint is removed in the latest versions of dbtemplates). Since one of the tests was defining two template with the same name, an exception was raised. Fixed this by giving different names to those two templates.